requirements.txt
requests
pyairtable
pycccedict
pypinyin
python-dotenv